Crate persistent

A set of middleware for sharing data between requests in the Iron framework.

Structs

Read
Middleware for data that persists between Requests with read-only capabilities.

State
Middleware for data that persists between requests with read and write capabilities.

Write
Middleware for data that persists between Requests for data which mostly needs to be written instead of read.

Enums

PersistentError
The type that can be returned by eval to indicate error.