Struct persistent::Read

pub struct Read<P: Key> {
    // some fields omitted
}

Middleware for data that persists between Requests with read-only capabilities.

The data is stored behind an Arc, so multiple threads can have concurrent, non-blocking access.

Read can be linked as BeforeMiddleware to add data to the Request extensions and it can be linked as an AfterMiddleware to add data to the Response extensions.

Read also implements Plugin, so the data stored within can be accessed through request.get::<Read<P>>() as an Arc<P::Value>.

Methods

impl<P: Key> Read<P> where P: Send + Sync

fn both(start: P) -> (Read<P>, Read<P>)

Construct a new pair of Read that can be passed directly to Chain::link.

The data is initialized with the passed-in value.

fn one(start: P) -> Read<P>

Construct a new Read that can be passed directly to Chain::link_before or Chain::link_after.

The data is initialized with the passed-in value.