Struct persistent::Write

pub struct Write<P: Key> {
    // some fields omitted
}

Middleware for data that persists between Requests for data which mostly needs to be written instead of read.

The data is stored behind a Mutex, so only one request at a time can access the data. This is more performant than State in the case where most uses of the data require a write lock.

Write can be linked as BeforeMiddleware to add data to the Request extensions and it can be linked as an AfterMiddleware to add data to the Response extensions.

Write also implements Plugin, so the data stored within can be accessed through request.get::<Write<P>>() as an Arc<Mutex<P::Value>>.

Methods

impl<P: Key> Write<P> where P: Send

fn both(start: P) -> (Write<P>, Write<P>)

Construct a new pair of Write that can be passed directly to Chain::link.

The data is initialized with the passed-in value.

fn one(start: P) -> Write<P>

Construct a new Write that can be passed directly to Chain::link_before or Chain::link_after.

The data is initialized with the passed-in value.